Restore nested comb helper calls
